# src/controllers/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402 Semantic Transaction Summary API",
    "version": "0.1.0",
    "description": "Semantic blockchain transaction decoder protected by x402 pay-per-call middleware.\n\nMain flow:\n1. Client calls `GET /summary?network=&tx_hash=` or `GET /summary/{network}/{tx_hash}`.\n2. If unpaid, API returns `402 Payment Required` with accepted payment requirements.\n3. Client pays via x402 and retries.\n4. API returns semantic summary JSON.\n",
    "guidance": "Use GET with query parameters or path parameters for `network` and `tx_hash`.\nTreat this as a paid endpoint via x402:\n- First call may return 402 Payment Required.\n- Pay using x402 (v2 header `Payment-Signature` or an x402 SDK) and retry the same URL.\n- Prefer public HTTPS origin (e.g. api.dappdojo.com) so resource URLs are consistent.\n"
  },
  "servers": [
    {
      "url": "https://api.dappdojo.com",
      "description": "Custom domain (optional)"
    },
    {
      "url": "https://x402-tx-summary.fly.dev",
      "description": "Fly deployment"
    }
  ],
  "tags": [
    {
      "name": "Summary"
    }
  ],
  "paths": {
    "/summary": {
      "get": {
        "tags": [
          "Summary"
        ],
        "summary": "Decode transaction into semantic intent (query parameters)",
        "description": "Decodes a transaction from the specified network into a normalized semantic structure.\n\nQuery: `network` (chain id) and `tx_hash`. This route is protected by x402 middleware and may return `402 Payment Required`\nuntil a valid payment header is provided.\n",
        "operationId": "getSummaryQuery",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "amount": "0.001",
            "currency": "USD"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "extensions": {
          "bazaar": {
            "schema": {
              "type": "object",
              "properties": {
                "input": {
                  "type": "object",
                  "properties": {
                    "queryParams": {
                      "type": "object",
                      "required": ["network", "tx_hash"],
                      "properties": {
                        "network": {
                          "type": "string",
                          "description": "Numeric EVM chain id.",
                          "enum": ["1", "8453", "84532"]
                        },
                        "tx_hash": {
                          "type": "string",
                          "description": "0x-prefixed 32-byte transaction hash.",
                          "pattern": "^0x[a-fA-F0-9]{64}$"
                        }
                      }
                    }
                  }
                },
                "output": {
                  "type": "object",
                  "properties": {
                    "example": {
                      "type": "object",
                      "properties": {
                        "schema_version": { "type": "string" },
                        "tx_hash": { "type": "string" },
                        "data": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-schema": {
          "input": {
            "type": "object",
            "required": [
              "network",
              "tx_hash"
            ],
            "properties": {
              "network": {
                "type": "string",
                "enum": [
                  "1",
                  "8453",
                  "84532"
                ]
              },
              "tx_hash": {
                "type": "string",
                "pattern": "^0x[a-fA-F0-9]{64}$"
              }
            }
          },
          "output": {
            "type": "object",
            "properties": {
              "schema_version": {
                "type": "string"
              },
              "tx_hash": {
                "type": "string"
              },
              "data": {
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "network",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Numeric EVM chain id.",
              "enum": ["1", "8453", "84532"]
            }
          },
          {
            "name": "tx_hash",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "0x-prefixed 32-byte transaction hash.",
              "pattern": "^0x[a-fA-F0-9]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decoded semantic transaction summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchedTxData"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (network or tx hash)",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "examples": {
                    "invalidNetwork": {
                      "value": "Invalid network"
                    },
                    "invalidTxHash": {
                      "value": "Invalid tx hash"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/X402PaymentRequiredV2"
                },
                "examples": {
                  "baseUsdc": {
                    "value": {
                      "resource": {
                        "url": "https://api.dappdojo.com/summary",
                        "description": "Semantic transaction decode result",
                        "mimeType": "application/json"
                      },
                      "accepts": [
                        {
                          "scheme": "exact",
                          "network": "eip155:8453",
                          "amount": "1000",
                          "payTo": "0x03ae0E3557A3e708aBEFAF0a8248D0e09CaA57E3",
                          "maxTimeoutSeconds": 300,
                          "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                          "extra": {
                            "assetTransferMethod": "eip3009",
                            "name": "USD Coin",
                            "version": "2"
                          }
                        }
                      ],
                      "error": "Verification failed: address_not_registered",
                      "x402Version": 2
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Transaction not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/summary/{network}/{tx_hash}": {
      "get": {
        "tags": ["Summary"],
        "summary": "Decode transaction into semantic intent (path parameters)",
        "description": "Same as `GET /summary` but `network` and `tx_hash` are path segments. Protected by x402; may return `402 Payment Required` until paid.\n",
        "operationId": "getSummaryPath",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "amount": "0.001",
            "currency": "USD"
          },
          "protocols": [{ "x402": {} }]
        },
        "extensions": {
          "bazaar": {
            "schema": {
              "type": "object",
              "properties": {
                "input": {
                  "type": "object",
                  "properties": {
                    "queryParams": {
                      "type": "object",
                      "required": ["network", "tx_hash"],
                      "properties": {
                        "network": {
                          "type": "string",
                          "description": "Numeric EVM chain id.",
                          "enum": ["1", "8453", "84532"]
                        },
                        "tx_hash": {
                          "type": "string",
                          "description": "0x-prefixed 32-byte transaction hash.",
                          "pattern": "^0x[a-fA-F0-9]{64}$"
                        }
                      }
                    }
                  }
                },
                "output": {
                  "type": "object",
                  "properties": {
                    "example": {
                      "type": "object",
                      "properties": {
                        "schema_version": { "type": "string" },
                        "tx_hash": { "type": "string" },
                        "data": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-schema": {
          "input": {
            "type": "object",
            "required": ["network", "tx_hash"],
            "properties": {
              "network": { "type": "string", "enum": ["1", "8453", "84532"] },
              "tx_hash": { "type": "string", "pattern": "^0x[a-fA-F0-9]{64}$" }
            }
          },
          "output": {
            "type": "object",
            "properties": {
              "schema_version": { "type": "string" },
              "tx_hash": { "type": "string" },
              "data": { "type": "object" }
            }
          }
        },
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["1", "8453", "84532"]
            }
          },
          {
            "name": "tx_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decoded semantic transaction summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FetchedTxData" }
              }
            }
          },
          "400": {
            "description": "Invalid input (network or tx hash)",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "examples": {
                    "invalidNetwork": { "value": "Invalid network" },
                    "invalidTxHash": { "value": "Invalid tx hash" }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/X402PaymentRequiredV2" },
                "examples": {
                  "baseUsdc": {
                    "value": {
                      "resource": {
                        "url": "https://api.dappdojo.com/summary/8453/0x47807d99d1748731e70eaf66da01ac822b845c179053cb652b0f08ba444b24a1",
                        "description": "Semantic transaction decode result",
                        "mimeType": "application/json"
                      },
                      "accepts": [
                        {
                          "scheme": "exact",
                          "network": "eip155:8453",
                          "amount": "1000",
                          "payTo": "0x03ae0E3557A3e708aBEFAF0a8248D0e09CaA57E3",
                          "maxTimeoutSeconds": 300,
                          "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                          "extra": {
                            "assetTransferMethod": "eip3009",
                            "name": "USD Coin",
                            "version": "2"
                          }
                        }
                      ],
                      "error": "Verification failed: address_not_registered",
                      "x402Version": 2
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "text/plain": {
                "schema": { "type": "string", "example": "Transaction not found" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "text/plain": { "schema": { "type": "string" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "PaymentSignatureHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Payment-Signature",
        "description": "x402 v2 payment payload header used for paid retry requests.\nClients usually generate this automatically using an x402-aware SDK.\n"
      }
    },
    "schemas": {
      "FetchedTxData": {
        "type": "object",
        "required": [
          "schema_version",
          "labels",
          "block_number",
          "block_hash",
          "tx_hash",
          "transaction_index",
          "data",
          "signer"
        ],
        "properties": {
          "schema_version": {
            "type": "string",
            "example": "0.1.0"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "erc20",
              "transfer",
              "stablecoin"
            ]
          },
          "block_number": {
            "type": "integer",
            "format": "uint64",
            "example": 26104298
          },
          "block_hash": {
            "type": "string",
            "example": "0x769909f41d6ea5fa4e5e007c940b19dbd66ca75b9d60208521da57ab3f7d0547"
          },
          "tx_hash": {
            "type": "string",
            "example": "0x47807d99d1748731e70eaf66da01ac822b845c179053cb652b0f08ba444b24a1"
          },
          "transaction_index": {
            "type": "integer",
            "format": "uint64",
            "example": 52
          },
          "signer": {
            "type": "string",
            "example": "0xE5Cff9Ac97633c46C73B92C85927e966f7d4EfaC"
          },
          "data": {
            "$ref": "#/components/schemas/DecodedTxData"
          }
        }
      },
      "DecodedTxData": {
        "type": "object",
        "required": [
          "chain",
          "tx_type",
          "nonce",
          "gas",
          "to",
          "actions"
        ],
        "properties": {
          "chain": {
            "$ref": "#/components/schemas/ChainInfo"
          },
          "tx_type": {
            "type": "string",
            "enum": [
              "legacy",
              "eip1559",
              "unknown"
            ],
            "example": "eip1559"
          },
          "nonce": {
            "type": "integer",
            "format": "uint64",
            "example": 234
          },
          "gas": {
            "$ref": "#/components/schemas/Gas"
          },
          "to": {
            "type": "string",
            "example": "0xfde4c96c8593536e31f229ea8f37b2ada2699bb2"
          },
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InputTxData"
            }
          }
        }
      },
      "ChainInfo": {
        "type": "object",
        "required": [
          "chain_id",
          "name",
          "native_asset"
        ],
        "properties": {
          "chain_id": {
            "type": "integer",
            "format": "uint64",
            "example": 8453
          },
          "name": {
            "type": "string",
            "example": "Base"
          },
          "native_asset": {
            "$ref": "#/components/schemas/NativeAsset"
          }
        }
      },
      "NativeAsset": {
        "type": "object",
        "required": [
          "symbol",
          "decimals"
        ],
        "properties": {
          "symbol": {
            "type": "string",
            "example": "ETH"
          },
          "decimals": {
            "type": "integer",
            "format": "uint32",
            "example": 18
          }
        }
      },
      "Gas": {
        "type": "object",
        "required": [
          "limit",
          "price",
          "effective_gas_price",
          "max_fee_wei",
          "max_fee_eth"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "uint64",
            "example": 52149
          },
          "price": {
            "type": "integer",
            "format": "uint64",
            "example": 11011442
          },
          "effective_gas_price": {
            "type": "integer",
            "format": "uint64",
            "example": 3840874
          },
          "max_fee_wei": {
            "type": "integer",
            "format": "uint64",
            "example": 574235688858
          },
          "max_fee_eth": {
            "type": "number",
            "format": "double",
            "example": 5.74235688858e-07
          }
        }
      },
      "InputTxData": {
        "type": "object",
        "required": [
          "type",
          "subtype",
          "intent",
          "summary",
          "participants",
          "assets_in",
          "assets_out",
          "direction"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "transfer"
          },
          "subtype": {
            "type": "string",
            "example": "erc20"
          },
          "intent": {
            "type": "string",
            "example": "send_money"
          },
          "summary": {
            "type": "string"
          },
          "participants": {
            "$ref": "#/components/schemas/Participants"
          },
          "assets_in": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenInfo"
            }
          },
          "assets_out": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenInfo"
            }
          },
          "protocol": {
            "type": "string",
            "nullable": true
          },
          "function": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FunctionInfo"
              }
            ],
            "nullable": true
          },
          "direction": {
            "type": "string",
            "enum": [
              "incoming",
              "outgoing"
            ]
          }
        }
      },
      "FunctionInfo": {
        "type": "object",
        "required": [
          "selector",
          "name"
        ],
        "properties": {
          "selector": {
            "type": "string",
            "example": "0xa9059cbb"
          },
          "name": {
            "type": "string",
            "example": "transfer"
          }
        }
      },
      "Participants": {
        "type": "object",
        "required": [
          "sender",
          "receiver"
        ],
        "properties": {
          "sender": {
            "type": "string",
            "example": "0xE5Cff9Ac97633c46C73B92C85927e966f7d4EfaC"
          },
          "receiver": {
            "type": "string",
            "example": "0x0fA50a1E47E9Af209D1D0D82F896B9353Ea72861"
          }
        }
      },
      "TokenInfo": {
        "type": "object",
        "required": [
          "name",
          "symbol",
          "decimals",
          "raw_amount",
          "amount"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Tether USD"
          },
          "symbol": {
            "type": "string",
            "example": "USDT"
          },
          "decimals": {
            "type": "integer",
            "format": "uint64",
            "example": 6
          },
          "raw_amount": {
            "type": "string",
            "example": "150049661202"
          },
          "amount": {
            "type": "string",
            "example": "150049.661202"
          },
          "token_address": {
            "type": "string",
            "nullable": true,
            "example": "0xfde4C96c8593536E31F229EA8f37b2ADa2699bb2"
          }
        }
      },
      "X402PaymentRequiredV2": {
        "type": "object",
        "required": [
          "accepts",
          "x402Version",
          "resource"
        ],
        "properties": {
          "accepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/X402AcceptV2"
            }
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "x402Version": {
            "type": "integer",
            "enum": [
              2
            ]
          },
          "resource": {
            "$ref": "#/components/schemas/X402ResourceInfo"
          }
        }
      },
      "X402AcceptV2": {
        "type": "object",
        "required": [
          "scheme",
          "payTo",
          "asset",
          "network",
          "amount",
          "maxTimeoutSeconds"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "example": "exact"
          },
          "payTo": {
            "type": "string",
            "example": "0x03ae0E3557A3e708aBEFAF0a8248D0e09CaA57E3"
          },
          "asset": {
            "type": "string",
            "example": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
          },
          "network": {
            "type": "string",
            "example": "eip155:8453"
          },
          "amount": {
            "type": "string",
            "example": "1000"
          },
          "maxTimeoutSeconds": {
            "type": "integer",
            "format": "uint32",
            "example": 300
          },
          "extra": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "X402ResourceInfo": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}